==> 1-DataSplit.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.metrics import accuracy_score, roc_curve, auc
from sklearn.model_selection import train_test_split
from sklearn.utils import shuffle
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier

data = pd.read_csv("ppi.csv",index_col=0)

data.describe()

for i,col in enumerate(data.columns):
    print(i,col)

data = data[data.sequence != 'X']
data.describe()

data.sequence=data.sequence.astype('category')
print(data.sequence.dtype)

len(list(data.columns))

edata=data
cate_features=[4]
enc = OneHotEncoder(handle_unknown='ignore')
column_transformer = ColumnTransformer(
[('encoder', enc, cate_features)],
remainder='passthrough',
verbose_feature_names_out=False
)
edata = column_transformer.fit_transform(edata)
encoded_df = pd.DataFrame(edata, columns=column_transformer.get_feature_names_out(list(data.columns)))
encoded_df.describe()


encoded_df

y=encoded_df.uniprot_id
y=y.drop_duplicates()
y

y1=y.sample(frac=1)

y1.iloc[:45]

test_prots1=list(y1.iloc[:45])
train_prots1=list(y1.iloc[45:])

len(test_prots1)


len(train_prots1)

test1 = encoded_df[encoded_df['uniprot_id'].isin(test_prots1)]
train1 = encoded_df[encoded_df['uniprot_id'].isin(train_prots1)]

test1.p_interface.value_counts()

encoded_df.p_interface.value_counts()

train1.to_csv('train1.csv', index=False)
test1.to_csv('test1.csv',index=False)

x_train1 = train1.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y_train1 = train1.p_interface
x_test1 = test1.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y_test1 = test1.p_interface

rf = RandomForestClassifier()
rf.fit(x_train1, y_train1)
rf_predictions = rf.predict(x_test1)
print("Accuracy of Random Forest:", accuracy_score(y_test1, rf_predictions))

y_rf_prob = rf.predict_proba(x_test1)[:, 1]
roc_rf = roc_curve(y_test1, y_rf_prob)
print("AUC of Random Forest:", auc(roc_rf[0], roc_rf[1]))

xgboost = XGBClassifier()
xgboost.fit(x_train1, y_train1)
xgb_predictions = xgboost.predict(x_test1)
print("Accuracy of XGBoost:", accuracy_score(y_test1, xgb_predictions))

y_xgb_prob = xgboost.predict_proba(x_test1)[:, 1]
roc_xgb = roc_curve(y_test1, y_xgb_prob)
print("AUC of XGBoost:", auc(roc_xgb[0], roc_xgb[1]))

import joblib
joblib.dump(rf,'rf1.pkl')
joblib.dump(xgboost,'xgb1.pkl')

cluster_data=data

cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['A','G','V']),1)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['I','L','F','P']),2)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['Y','M','T','S']),3)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['H','N','Q','W']),4)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['R','K']),5)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"].isin(['D','E']),6)
cluster_data["sequence"] = cluster_data["sequence"].mask(cluster_data["sequence"]=='C',7)

cluster_data.sequence

cate_features=[4]
enc = OneHotEncoder(handle_unknown='ignore')
column_transformer = ColumnTransformer(
[('encoder', enc, cate_features)],
remainder='passthrough',
verbose_feature_names_out=False
)
cluster_data = column_transformer.fit_transform(cluster_data)
encoded_df = pd.DataFrame(cluster_data, columns=column_transformer.get_feature_names_out(list(data.columns)))
encoded_df.describe()

train1 = pd.read_csv("train1.csv")
test1 = pd.read_csv("test1.csv")

test_prots2=list(test1.uniprot_id.drop_duplicates())
train_prots2=list(train1.uniprot_id.drop_duplicates())

test2 = encoded_df[encoded_df['uniprot_id'].isin(test_prots2)]
train2 = encoded_df[encoded_df['uniprot_id'].isin(train_prots2)]

test2.sequence_1.dtype

test2.p_interface.value_counts()

train2.to_csv('train2.csv', index=False)
test2.to_csv('test2.csv',index=False)

ordinal_data=data

ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['A','G','V']),1.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['I','L','F','P']),2.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['Y','M','T','S']),3.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['H','N','Q','W']),4.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['R','K']),5.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"].isin(['D','E']),6.0)
ordinal_data["sequence"] = ordinal_data["sequence"].mask(ordinal_data["sequence"]=='C',7.0)

ordinal_data.sequence.astype('float64')

test3 = ordinal_data[ordinal_data['uniprot_id'].isin(test_prots2)]
train3 = ordinal_data[ordinal_data['uniprot_id'].isin(train_prots2)]

test3.p_interface.value_counts()

train3.to_csv('train3.csv', index=False)
test3.to_csv('test3.csv',index=False)

removed_data=data

removed_data=removed_data.drop('sequence',axis=1)

test4 = removed_data[removed_data['uniprot_id'].isin(test_prots2)]
train4 = removed_data[removed_data['uniprot_id'].isin(train_prots2)]

test4.p_interface.value_counts()

train4.to_csv('train4.csv', index=False)
test4.to_csv('test4.csv',index=False)
