==> 2-DataAugmentation.ipynb <==
import pandas as pd
import numpy as np
from imblearn.under_sampling import RandomUnderSampler
from collections import Counter
from imblearn.under_sampling import NearMiss
from imblearn.over_sampling import SMOTENC

train1 = pd.read_csv("train1.csv")

x = train1.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y = train1.p_interface

print('Original dataset shape %s' % Counter(y))

rus = RandomUnderSampler(random_state=36)

rus_x,rus_y = rus.fit_resample(x, y)

print('Resampled dataset shape %s' % Counter(rus_y))

rus_x.to_csv('rus_x.csv', index=False)
rus_y.to_csv('rus_y.csv',index=False)

NM = NearMiss(version=3)

nm_x,nm_y = NM.fit_resample(x,y)

print('Resampled dataset shape %s' % Counter(nm_y))

nm_y

x

import pickle
with open("nearmiss.pickle", "wb") as file:
    pickle.dump((nm_x, nm_y), file)

with open("nearmiss.pickle", "rb") as file:
    nm_x, nm_y = pickle.load(file)

smote = SMOTENC(random_state=42, categorical_features=list(range(0,21)))

smote_x,smote_y = smote.fit_resample(x,y)

print('Resampled dataset shape %s' % Counter(smote_y))

with open("smotenc.pickle", "wb") as file:
    pickle.dump((smote_x, smote_y), file)

from imblearn.combine import SMOTEENN
sme = SMOTEENN(random_state=42,smote=smote)

sme_x,sme_y = sme.fit_resample(x,y)

print('Resampled dataset shape %s' % Counter(smote_y))

with open("smoteenn.pickle", "wb") as file:
    pickle.dump((sme_x, sme_y), file)

train2 = pd.read_csv("train2.csv")

x = train2.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y = train2.p_interface

print('Original dataset shape %s' % Counter(y))

rus = RandomUnderSampler(random_state=36)

rus_x,rus_y = rus.fit_resample(x, y)

print('Resampled dataset shape %s' % Counter(rus_y))

import pickle
with open("RUS2.pickle", "wb") as file:
    pickle.dump((rus_x, rus_y), file)

train3 = pd.read_csv("train3.csv")

x = train3.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y = train3.p_interface

rus = RandomUnderSampler(random_state=36)

rus_x,rus_y = rus.fit_resample(x, y)

print('Resampled dataset shape %s' % Counter(rus_y))

import pickle
with open("RUS3.pickle", "wb") as file:
    pickle.dump((rus_x, rus_y), file)

train4 = pd.read_csv("train4.csv")

x = train4.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y = train4.p_interface

rus = RandomUnderSampler(random_state=36)
rus_x,rus_y = rus.fit_resample(x, y)
print('Resampled dataset shape %s' % Counter(rus_y))

import pickle
with open("RUS4.pickle", "wb") as file:
    pickle.dump((rus_x, rus_y), file)
