==> 3-tuning.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import RandomizedSearchCV
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score, roc_curve, auc
import matplotlib.pyplot as plt
%matplotlib inline
import pickle

x = pd.read_csv("rus_x.csv")
y = pd.read_csv("rus_y.csv")

y

y=y.to_numpy().ravel()

y

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,300,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(x, y)

search.best_params_

search.best_score_

cv_results = pd.DataFrame(search.cv_results_)

cv_results.params[14]

import pickle
with open("nearmiss.pickle", "rb") as file:
    nm_x, nm_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,300,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(nm_x, nm_y)

search.best_params_

search.best_score_

with open("smotenc.pickle", "rb") as file:
    smote_x, smote_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,300,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(smote_x, smote_y)

search.best_params_

search.best_score_

with open("smoteenn.pickle", "rb") as file:
    sme_x, sme_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(140,280,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(sme_x, sme_y)

search.best_params_

search.best_score_

test1 = pd.read_csv("test1.csv")
x_test1 = test1.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y_test1 = test1.p_interface

train1 = pd.read_csv("train1.csv")
x_train1 = train1.drop(['aa_ProtPosition', 'domain','p_interface','uniprot_id','Rlength'], axis=1)
y_train1 = train1.p_interface

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,300,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=10, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(x_train1, y_train1)

search.best_params_

search.best_score_

bst = XGBClassifier()
distributions = {"max_depth": [4,6,8,10],
                 "n_estimators": np.arange(50, 200, 20)}
bst_clf = RandomizedSearchCV(bst, param_distributions=distributions,n_iter=10, random_state=0,verbose=1,scoring='f1')
search = bst_clf.fit(x, y)

print(search.best_params_)
print(search.best_score_)

bst = XGBClassifier()
distributions = {"max_depth": [4,6,8,10],
                 "n_estimators": np.arange(50, 200, 20)}
bst_clf = RandomizedSearchCV(bst, param_distributions=distributions,n_iter=10, random_state=0,verbose=1,scoring='f1')
search = bst_clf.fit(x_train1, y_train1)

print(search.best_params_)
print(search.best_score_)

bst = XGBClassifier()
distributions = {"max_depth": [4,6,8,10],
                 "n_estimators": np.arange(50, 200, 20)}
bst_clf = RandomizedSearchCV(bst, param_distributions=distributions,n_iter=10, random_state=0,verbose=1,scoring='f1')
search = bst_clf.fit(smote_x, smote_y)

print(search.best_params_)
print(search.best_score_)

with open("RUS2.pickle", "rb") as file:
    rus2_x, rus2_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,300,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(rus2_x, rus2_y)

print(search.best_params_)
print(search.best_score_)

with open("RUS2.pickle", "rb") as file:
    rus2_x, rus2_y = pickle.load(file)

rus2_x = rus2_x.drop(['normalized_length'],axis=1)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,220,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=10, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(rus2_x, rus2_y)

print(search.best_params_)
print(search.best_score_)

with open("RUS3.pickle", "rb") as file:
    rus3_x, rus3_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,220,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(rus3_x, rus3_y)

print(search.best_params_)
print(search.best_score_)

with open("RUS4.pickle", "rb") as file:
    rus4_x, rus4_y = pickle.load(file)

rf = RandomForestClassifier(random_state=0,n_jobs=-1)
distributions = {"n_estimators": np.arange(100,220,20),
                 "criterion": ["gini"],
                 "max_depth": np.arange(10,100,10)}
rf_clf = RandomizedSearchCV(rf, param_distributions=distributions,n_iter=20, random_state=0,verbose=1,scoring='f1')
search = rf_clf.fit(rus4_x, rus4_y)

print(search.best_params_)
print(search.best_score_)
